export_third_party.py takes in path commandline arg

+++ export_third_party.py
@@ -8,15 +8,22 @@ from __future__ import print_function
 import sys, os, time
 from os.path import dirname
 from os.path import join
+import argparse
 
-path = '/home/wdai/lib/hotbox'
+parser = argparse.ArgumentParser()
+parser.add_argument("--install_path")
+args = parser.parse_args()
+if not args.install_path:
+  print('--install_path must be specified (installed to install_path/lib'
+    + ', install_path/include/hotbox.)')
+  sys.exit(1)
 
-lib_path = join(path, 'lib')
+lib_path = join(args.install_path, 'lib')
 print('lib_path:', lib_path)
 os.system('mkdir -p %s' % lib_path)
 os.system('cp build/lib/libhotbox.so %s' % lib_path)
 
-include_path = join(path, 'include', 'hotbox')
+include_path = join(args.install_path, 'include', 'hotbox')
 print('include path:', include_path)
 os.system('rm -r %s' % include_path)
 os.system('mkdir -p %s' % include_path)
